syskit: update for the cleanup on Engine in Syskit 2
