+++ ruby/lib/transformer/syskit/extensions.rb
@@ -4,7 +4,7 @@ require 'transformer/syskit/extensions/component_model'
 require 'transformer/syskit/extensions/component'
 require 'transformer/syskit/extensions/composition'
 require 'transformer/syskit/extensions/device'
-require 'transformer/syskit/extensions/engine'
+require 'transformer/syskit/extensions/system_network_generator'
 require 'transformer/syskit/extensions/graphviz'
 require 'transformer/syskit/extensions/instance_requirements'
 require 'transformer/syskit/extensions/plan'
